% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SDE_data.R
\name{get_SDE_data}
\alias{get_SDE_data}
\title{Function for downloading geodata from kortforsyningen.dk}
\usage{
get_SDE_data(
  ftp_folder = ftp_folder,
  file_name = file_name,
  out_folder = out_folder,
  userpwd = SDE_user
)
}
\arguments{
\item{ftp_folder}{Name of the ftp folder}

\item{file_name}{Name of the file we would like to download}

\item{out_folder}{Name of the local folder where the data will be saved}

\item{SDE_user}{Username and Password (format = USERNAME:PASSWORD)}
}
\value{
Data from \url{https://download.kortforsyningen.dk/content/geodataprodukter}
}
\description{
Function for downloading geodata from kortforsyningen.dk
}
\examples{

# Set Username and Passwordm in .Renviron (usethis::edit_r_environ())
  id <- Sys.getenv("kortforsyningen_id")
  password <- Sys.getenv("kortforsyningen_pwd")
  SDE_user <- paste(id, password, sep = ":" )

# Get data
  get_SDE_data(ftp_folder = "CORINE",
    file_name  = "DK_CORINE_SHP_UTM32-WGS84.zip",
    out_folder = "data/CORINE")

}
