% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dangeo_get_data.R
\name{dangeo_get_data}
\alias{dangeo_get_data}
\title{Function for downloading geodata from kortforsyningen.dk}
\usage{
dangeo_get_data(
  ftp_folder = NULL,
  zip_name = NULL,
  out_folder = rappdirs::user_cache_dir(),
  file_name = NULL,
  userpwd = set_user
)
}
\arguments{
\item{ftp_folder}{Name of the ftp folder}

\item{zip_name}{Name of the zip file we would like to download..}

\item{out_folder}{Name of the local folder where the data will be saved (by default it is set to the results of rappdirs::user_cache_dir()).}

\item{file_name}{Name of the file (e.g. *.shp) we would like to load into R.}

\item{userpwd}{Username and Password (format = USERNAME:PASSWORD).}
}
\value{
Data from \url{https://download.kortforsyningen.dk/content/geodataprodukter}.
}
\description{
Function for downloading geodata from kortforsyningen.dk
}
\examples{

# Set Username and Passwordm in .Renviron (usethis::edit_r_environ())
  id <- Sys.getenv("kortforsyningen_id")
  password <- Sys.getenv("kortforsyningen_pwd")
  set_user <- paste(id, password, sep = ":" )

# Get data
dangeo_get_data(ftp_folder = "CORINE",
                zip_name   = "DK_CORINE_SHP_UTM32-WGS84.zip",
                out_folder = "DK_CORINE_SHP_UTM32-WGS84",
                file_name  = "CLC12_DK.shp")

}
